Updated Projects JSON & Remove outdated nested projects
- Outdated Architecture, can now utilize Github Submodules
