+++ src/assets/styles/App.scss
@@ -9,8 +9,5 @@
 @import './modules/footer';
 @import './modules/modal';
 
-// Animal Crossing Styling
-@import './actunes/index';
-
 // JavaScript Shorts Styling
 @import './widgets/index';
