--- src/assets/styles/actunes/_index.scss
@@ -1,89 +1,0 @@
-.ACTunes_body {
-  background-color: $primaryBlue;
-
-  header img {
-    margin-top: 1rem;
-    width: 20rem;
-  }
-  .weatherIcon {
-    background-color: ghostwhite;
-    width: 2rem;
-    padding: 5px;
-    cursor: pointer;
-    margin: 0.5em;
-  }
-
-  .clock {
-    position: relative;
-    text-align: center;
-    width: 20rem;
-    margin: auto;
-
-    img {
-      width: 100%;
-    }
-
-    &_time {
-      position: absolute;
-      bottom: 12%;
-      left: 14%;
-      color: black;
-      font-size: 3.5rem;
-      transform: rotate(24deg);
-    }
-  }
-
-  #player {
-    background: whitesmoke;
-    width: 25rem;
-    margin: 0;
-  }
-
-  #weatherForm {
-    input {
-      color: white;
-      background-color: rgba(255, 255, 255, 0.1);
-      width: 20rem;
-      text-align: center;
-      display: block;
-      margin: auto;
-    }
-
-    .btn {
-      margin-top: 1rem;
-    }
-  }
-
-  .geo {
-    color: black;
-    padding: 0;
-  }
-
-  footer,
-  .timeline {
-    margin-top: 1rem;
-  }
-
-  footer {
-    margin-bottom: 3rem;
-  }
-
-  .downloads {
-    padding: 0;
-    select {
-      display: block;
-    }
-    form {
-      display: flex;
-    }
-
-    &_album {
-      display: flex;
-      flex-wrap: wrap;
-      justify-content: space-around;
-    }
-    &_title {
-      flex-basis: 100%;
-    }
-  }
-}
