--- src/components/addons/gameboyApp.html
@@ -1,35 +1,0 @@
-<div className="gameboy-App">
-  <a href="https://github.com/baumannzone/gameboy-css">
-    <img
-      class="gh-img"
-      src="https://s3.amazonaws.com/github/ribbons/forkme_right_red_aa0000.png"
-      alt="Fork me on GitHub"
-    />
-  </a>
-  <input class="btn-on" type="button" value="Encender" />
-  <input class="btn-off btn-hide" type="button" value="Apagar" />
-  <audio src="sound/gameboy-sound.mp3"></audio>
-  <div class="gameboy">
-    <div class="screen-cont">
-      <div class="power"></div>
-      <div class="screen">
-        <div class="header">DOT MATRIX WITH STEREO SOUND</div>
-        <div class="animated-text">
-          Nintendo
-          <div class="copy">®</div>
-        </div>
-      </div>
-    </div>
-    <div class="controls-cont">
-      <div class="btn-direction">
-        <div class="vertical"></div>
-        <div class="horizontal"></div>
-      </div>
-      <div class="btn-AB"></div>
-      <div class="btn-start-select"></div>
-    </div>
-    <div class="speakers"></div>
-    <div class="on-off">< off-on ></div>
-    <div class="phones">phones</div>
-  </div>
-</div>
