--- src/components/projects/_actunes/_audioData.js
@@ -1,268 +1,0 @@
-const audioData = {
-	Alt: 'Complete OST collection that plays hourly while playing, Animal crossing. Only includes the 3 main games, excludes spinoff / Minigames',
-
-	Original: {
-		Normal: {
-			Midnight: '1LijLhyTtyejpLu-JfY32_fu1VjRgno3h',
-			'1AM': '1cHMvGMfrrSZUt1VeQFCogZC3b75YZfFp',
-			'2AM': '1ppSPssm3yRnagw_j6IZonbtJG5AL0fEb',
-			'3AM': '1n-8T-nNImwJYMWhN1nP_GFOSHZFjPo3r',
-			'4AM': '1FBDvf9jlcfVtTWMt98iI8HyNUDBzpMf5',
-			'5AM': '1K8VNvMCMnTZxgXDPbWY7BJBMUQQIhtHQ',
-			'6AM': '1cHRW5bPar5Dg4mrdAvjh54T2vqu_Db7F',
-			'7AM': '1ExXrhphRSUgWvulWhVgliMr2DkT_rcMl',
-			'8AM': '1sFbMXCUsKAkk6FXIzkG-IZAdVBwRr38y',
-			'9AM': '130V_x37KNLxzze78rk1bVpeH1Z21GVRM',
-			'10AM': '1eYkDSgiMj0CUTwHdd1yXkSn-wrZc55Xt',
-			'11AM': '1zFhApGq62WXIbUnxE81Z7klQ8VdGLZH2',
-			Noon: '1nbBZoeuaeJM0UPX4hNM3BCaPdiGFWUL2',
-			'1PM': '1wQZSezZloN3AVQGFqSi_f_3Aug_4bMeP',
-			'2PM': '1ALr9t7mCZLb1IoBa8isstpjHvp1IznEy',
-			'3PM': '1-isAGWi15-65mmNSvHAnj6D5ovz9O0r3',
-			'4PM': '1g2EP4heQuA5IxMsV2TMwOi5-GpizcZPH',
-			'5PM': '1NGdNr5eFQX7RT9BE2LbjqspI9XHf_RwU',
-			'6PM': '1-F6UtPy8GshmDyPFcHHq5Fc-CYZizGHB',
-			'7PM': '1TXMkfZaKvnCHwhVCf-d_1hdZgn2L168W',
-			'8PM': '1dd1gry5_ynz2omju3qTfjt0tRe_W6pY7',
-			'9PM': '1WWrwws20m4rXiL6IzNY0WCJpz-RYLm2S',
-			'10PM': '1CF2tympjq3Sjc5Rh97jHZAXr-s7xQm9f',
-			'11PM': '1R4S0NbFORYe4OrP5iShm7Npbh9fhtYhw',
-		},
-		Raining: {
-			Original: '1t_Ty6xJaqxqOp-wXGykiaZsZEALKU68y',
-			Midnight: '1pLbzfagJkrCBi6pMCOoKrm-WIP2vBWrG',
-			'1AM': '1-XptrJwvWsRwl_36x76tpvfeyk_bwvt_',
-			'2AM': '18cy-gKf7IoA_UOfm1Hu0ei-euseoQ89L',
-			'3AM': '1aekRdNtIMFczwzdCLlXK4PXnQiIHXF9V',
-			'4AM': '11_OMev6qjm3UTrdj20WK_yhuiLZJRg3B',
-			'5AM': '1IOrYF5hMm75yQy8iZWen-xFcs6bZytJv',
-			'6AM': '1_7VnexZlCxi264vv-Les3TO7fA11aFm5',
-			'7AM': '1UKpJ0k4FtmnLfEijsYgXsMD1g_SdoG6f',
-			'8AM': '1drISjEBU2Kzu9itAVFea7sDblWzxRZCa',
-			'9AM': '1rg3czfLGxxN-Wj6BSRaVa-RSB6AvO7lT',
-			'10AM': '1P5gEML96Ht9C6KLZwbd1zw3B1bvIIApC',
-			'11AM': '1QFJKDY4UArpP4o8bbjYZ9Ix5elrhqPd7',
-			Noon: '1qyu7vwUzkoivR81-ehqYeXqZFf8-otjd',
-			'1PM': '1sYwk7xkKcfcI0fZrIwrGAGUK2j98mKNJ',
-			'2PM': '1RqWXSZ5r3D-bztVPCDiZhrnNt9ts7ltL',
-			'3PM': '1I9yq5HeoSlvB75bz-tU1c99YUvPLBmYi',
-			'4PM': '1Hvcd78dVen2dNcJF7M95yTjYDgYmIWck',
-			'5PM': '1--IPgyUzyR_TrSFtNVTjCM5UH5azgV40',
-			'6PM': '1W3276_FPNBkS-gOC8yy05s5d2_51OLZY',
-			'7PM': '1OKSJuUVMlnCiJdThlG2Hkpr6sq_Cw7mE',
-			'8PM': '1OwiHIIfQQIQ5x2buubzmou3CJ0C_UeF3',
-			'9PM': '1yuQT3KkT1ifwGjlSxIRyYRMPXBYBNG7x',
-			'10PM': '17QZWswc8yyTQlvVEk-1oYyPpvRCKauHq',
-			'11PM': '1TjShan_W-5M5NSEW3wx2Udbujr6RxLm5',
-		},
-		Winter: {
-			Midnight: '1N0awi9KypMEEiUW7hkhgo3DjvRfy3qEC',
-			'1AM': '1Hmu5fFZV8iDPPydKRXlVlHCiQGQDptcR',
-			'2AM': '136W5r8epoRUAPOm8mvUaj2olH-gvMBTC',
-			'3AM': '1n6rrU8E17AcvUwX0GDjzJpKY-bojeCx7',
-			'4AM': '1DTiWVvgHz0KZLod8F2Nz7ag29Y2ia4P1',
-			'5AM': '1MCdGaFppVB2f8kQxEmXYasKvGwgd9Erm',
-			'6AM': '1llFTML2f3TTxa7w12rGZmBK-AyUHjnWZ',
-			'7AM': '1T7Bcqr5ssBb6CpDkFGp3p_FOPxvANH43',
-			'8AM': '1d3oIdjg4W7qKz_5hXnvoOgDFZn9BcO7-',
-			'9AM': '15ry3l5bBJMd5vvWeezW0m2-vmO7KyMPA',
-			'10AM': '1vjehiZzjcq_8TdOcv3xSHouvIvxq153X',
-			'11AM': '1uby-iAFncK9aTylDTIJS1AQwJiKTfizm',
-			Noon: '1n2v0ZRs79A-qw4Wdh_hQeJdpiQu5lUs0',
-			'1PM': '1kCGYc-Es7JKhFAa41QnxDATaDQtGoRF0',
-			'2PM': '1Z5n1pVVkWC9qcCKS9bHJpMHVHFrAPIhu',
-			'3PM': '1hpfQ5eDjhHZySaGgVE8XhUEgad3kdBtP',
-			'4PM': '1JWDXf5Pochgruk6bRTriljpPWRchzrAZ',
-			'5PM': '14ad8xDFDrTBy7S2X2bOSnFnYeQQskxyl',
-			'6PM': '1pmz6u359-BnuNIdrR23SuWKYe-ro1moX',
-			'7PM': '1uL1oG0C3dLDLlsmnffbVJlLbKjPiKEGz',
-			'8PM': '1TxRfUYop-i6SyL0ORVFqDUB7mR8KMDkg',
-			'9PM': '1BLlwJ5Zx2s1rp-HXGbZ_xL9pTnBtp8sV',
-			'10PM': '17yrDxogKpYOIyRtJgmKyIKiShJLG4VQu',
-			'11PM': '1gyTYnGRMUQ3Hbatokyb906KDIYO5SAcY',
-		},
-		Reference:
-			'https://downloads.khinsider.com/game-soundtracks/album/animal-crossing-full-gamerip',
-	},
-
-	NewLeaf: {
-		Normal: {
-			Midnight: '1TLRaOTlyncVJSxbqLjRfbBZ-Mif5-NCX',
-			'1AM': '19VAAxGpR9M6BZWxu-s9KCDzMLH-QNcdM',
-			'2AM': '196IuUGDPNtxjKOz6yDM7xJqWV3pR7SGF',
-			'3AM': '1j1_oMWXZmae4eT-hvLr2J7P4mRe2yjum',
-			'4AM': '1NJ-Sn5IrI4pSQk3n64hCe7iVxurYCLdF',
-			'5AM': '1F2c_rQzR98rtHRTsmrc8oDpRXwhHsX7o',
-			'6AM': '1CWcRRMDauP4Xawjr1JEo_pOOaYVHOwSR',
-			'7AM': '1TlByQAN353yGMPbatu0_qum9ajUVvNFj',
-			'8AM': '1UNHeX8jglSHcPJwfv8xXVTvPUcobolXe',
-			'9AM': '1ZC2M2FH-UpBvXrtWZfXcCj0RS12w7wle',
-			'10AM': '1f3r4lGLMXxKNRYrINPxGAS8m2txLzRMQ',
-			'11AM': '1VbMxWRu4fu0oyj32feIsMKrkSeW8LSQN',
-			Noon: '1msAU1ZDwFleY05Az6HXvbKLYrUKTXF8N',
-			'1PM': '12Q9g750WPWx-rMgCpGt1nAfItXYQe9al',
-			'2PM': '1n-fD7gHKyfeI49dNdXyTArUGGDefq9TS',
-			'3PM': '1wlP9KzedRFqTOdMEVRm5EFtIshrMBQaJ',
-			'4PM': '1SWJwmCJn087uCbwgQ1DAG83SD2bEVZK5',
-			'5PM': '13RPTvmV8tzwR8YkEkpPg2jrwDbZta0wr',
-			'6PM': '1VBVwprfHQbeA8_E_qZ1ZiRJMTxvtJUeh',
-			'7PM': '1Nolb_kBv4uXLrd_kUxU75SvCM6OJHZsp',
-			'8PM': '16Haq3v3iqgdXS4mDtnSICyaT7A8teNGJ',
-			'9PM': '1mWEFE1LxGmWAZlWGOvW_u-YOmKMes2rW',
-			'10PM': '1UyfOsdsBjdGAbgO8nkqq1ISRBfhCNoa0',
-			'11PM': '1bpm3YIunUecmyoiK2EEy-bOxg-P2oKHk',
-		},
-		Raining: {
-			Midnight: '1Nplue-GQb-cI9kxGeCDDcNntfTzALj35',
-			'1AM': '1lR5ARS66TnCAN--iA_6IPHYVscX4IFf6',
-			'2AM': '1juBv7u8AmvOp3cgOBJ21GTzYY-BjzEKe',
-			'3AM': '11XRDcGczLRBItW8EigzBO_vfHsqp6vbA',
-			'4AM': '1p8ZHijKmoOUMTyneZksrEWMUxjvkOKkl',
-			'5AM': '1bGJbN_NSlKVFUn26Ds-aoZeAz8ER3r0R',
-			'6AM': '1fbaK90Bp_kSH7YIy6wwVQ-OrmipLUqNj',
-			'7AM': '1lEUSEu-KjK-1bebBSys799wE6h0tji1d',
-			'8AM': '1DmbHNZpICrcp57jjaCfauAjqZUv6lIMI',
-			'9AM': '1Pp49ohi8ArD9Mv1cGKVyVQ7qqPvYehdZ',
-			'10AM': '1zYyMTuC8lBurFbl7rIlf23FwGptr-QeG',
-			'11AM': '1E11xYOsClkef-rM-IqYXOz-dskHyf_4S',
-			Noon: '1as5VpRhxoSGxuVZ-07UtxjivN7eHJBy0',
-			'1PM': '1l9JbF04nDgYZdZAhVdRAO2wPqu3fjDpy',
-			'2PM': '1RtZER3fjpC0ztBte3lxLiZetH4MODolu',
-			'3PM': '1attjIWRn4a1DuRF9ccMUCAdmL0DEBdbQ',
-			'4PM': '1V2gXl4Y2oQOGxTRgGoDqMbBhoz_qqnAm',
-			'5PM': '1Q7boXhYPXPOm_eDFpqkSK2iGPlyrgA2b',
-			'6PM': '1dLCnW0oa2Rr_nSj91dfuuIR9n3X3O4w6',
-			'7PM': '1tozl_IutLMr3apaz_ZznzmyLpUgCbMrY',
-			'8PM': '18Kkf1wXnId5g_KEsLiusiKONgy9usViK',
-			'9PM': '1pePm-PayPvSKLrQjjEMlRlPp2Lzf5kLM',
-			'10PM': '14RbFBWKiS969ZZFvvZudxvRq2EF-jSVq',
-			'11PM': '1udfAKi3wUzGSyRz18ToV92336bP-lbsj',
-		},
-		Winter: {
-			Midnight: '1pHw30Zeq7WeTK-f10JIuVMBa7uezWe1S',
-			'1AM': '1zhKpqt4_Cbw8cSVRgH2UXYIVxq6kWHhW',
-			'2AM': '1qqrvDm_BFVmgEx8L5aLQyfBJ0fmaSuNn',
-			'3AM': '1oEHad88PHJCpvWSGUovDJui8LRUy3aLS',
-			'4AM': '16T2Hwni5KWGtVe2QV74_GThZ1magAAOq',
-			'5AM': '1YgadZmi7q3yuFLBj3CqvSzt5fdKOVoTC',
-			'6AM': '1_Mx_BADbTgSSHXiD0eG-kFPMEBMJl5eo',
-			'7AM': '1_Cpet8haQjiFZcfU4rjv_R8YcLPOTvIV',
-			'8AM': '1WAn1uX29eLIwjiN3gCg7ESUnYopfTUDU',
-			'9AM': '1FKYsak25vJ4nQsbCFlCnH09nTKk3E-8x',
-			'10AM': '1SBXfHwyfaY6oIT7c-Ni1z1Wvp55AwasX',
-			'11AM': '1ferUzIZPahlqyKq88NKhHI2K7EUv14I-',
-			Noon: '1SnKChRzavzNzGOITrGkd3RgkYSk-zx9p',
-			'1PM': '1S0KQtRXiroeolSSJbK1oBBVJu91IswHm',
-			'2PM': '1-syR6QtkrLWTPUDBajVX7s_egyDDuPFH',
-			'3PM': '1LpahBgPehZAIhMcFW-lnP8gIPq2Im6fX',
-			'4PM': '1BUxQ3INNorYI2WguDWg7M7o8bYVCrtwc',
-			'5PM': '1vQ1RkL_yvCntV-53xqKYzUS33UfFQB2f',
-			'6PM': '1uB4_9BXOBbLywHOm6pMitDgSiEA7iARP',
-			'7PM': '1Nk3KAs2lhWwOENWMkHh2_92umVOJ1RR1',
-			'8PM': '1Iwuu7mT2NK1GzdrZ_9TTVPPTlYk9DWQx',
-			'9PM': '12QTfKSr8kf13uy95BPCVYe1501ocKG39',
-			'10PM': '1i_YGTLV7COomXT81bhwt5hwNl7sWGKDE',
-			'11PM': '1fey0X5MKnAtns55dca2S2zFKn4Ku8efg',
-		},
-		Reference: 'https://downloads.khinsider.com/game-soundtracks/album/animal-crossing-new-leaf',
-	},
-
-	CityFolk: {
-		Normal: {
-			Midnight: '1fAGRIHSqwCAQOekenqa63OvxLI9518zE',
-			'1AM': '1qo5aDVbFwXbNsC-A1CjlDH3r-Vd49OHd',
-			'2AM': '1nRKbWSzEVnQOSDMPFGf0Rak9kRM6SiQI',
-			'3AM': '14qnFypyZtVVm03M7VaKKRNMOw9dd3ErZ',
-			'4AM': '1rr39wP6qnahLBLYru-2ILxyCdjUwOJ1J',
-			'5AM': '1rR-cxqTQnXvLg3uRNALyqvaStkm9WRLu',
-			'6AM': '15YYG5bmB1zueWnpjxX3ZUqdZdQ4tO-lS',
-			'7AM': '1JoyB-RmaDY2fOjuDPkGFiL6VXOQr_Vkq',
-			'8AM': '1hc2FI5Bdv7pji7URVKDvb9gJuBDBGlIx',
-			'9AM': '1qedGzPB--qCIS0RJQ5KvfDM1JLwhu8C0',
-			'10AM': '1q-vAXtYGnuw37NCFq020JBuiQE_nb2_a',
-			'11AM': '1BR9XAcIcRbZAe5l-tSKu4OdKeBzXJU5D',
-			Noon: '19EtKv9rufdbTtPnNhX9FICH0C-1jNKb2',
-			'1PM': '1j7GVfCy6GdvWjIdulGyNj_Ye107ZCyWV',
-			'2PM': '1suyU6HLRtV5H-elJX1R7ASWk7wBiOlzB',
-			'3PM': '1oNgNAbVzmkY-Z58rHcMxAE8JxK7iq9qR',
-			'4PM': '13niHm-7OMW0K7uRi4LPB8JJUwfiLQ-XR',
-			'5PM': '1p92bVdwYayHNwOYUu773NsWKPba9U2EQ',
-			'6PM': '1MvLVQLtym4BqNi0WyfjhzcWdkiB3l7Ky',
-			'7PM': '1RGJbLxIlzOIZpG0mC99-afc3ESAfVl2e',
-			'8PM': '1rBrynzzTaM3R1_57hTg0D6oZfGgLI_4S',
-			'9PM': '1akgFTzNpYvmD5YRxFtSv_sfKhqCQITIf',
-			'10PM': '1y7MWfl4RPzTqlyCY5R3HbJX1tjbJO66X',
-			'11PM': '1mZqLU6W03j0VOQGDXsw3E1pKM4dTiW01',
-		},
-		Raining: {
-			Midnight: '1SrhR7mWDBJGZqmqk8yvp52ApVoWb8geE',
-			'1AM': '1Lqr0jearEicWo1u1Rt8DtiwKhPR3_kf_',
-			'2AM': '1ZhM6gt76lioJvf2RxvGvvWQoQEQYwQJf',
-			'3AM': '1qtvAreSG57x5hh4Dyrp6xVejw4-aDmZc',
-			'4AM': '1vb-ALvQCrdS0XPv8ZNYEMP53tx-OtXGB',
-			'5AM': '1EBcH6sKxVUGm0JVzsTMonWD_oIctIPxT',
-			'6AM': '1csv9ULI7QhR5VyXro1oEK7ZrTnHgTX5X',
-			'7AM': '1V4mmlL7LZnywYitUX72EkOEQEXziG7rx',
-			'8AM': '1E1BJKF65UN-4H9NoZG9M-EP4joHf7woa',
-			'9AM': '14QfFv4hEMcW9TjBAJs0ZmgRfufg-xIrp',
-			'10AM': '1m9uwVUNnYMBrKWq2gr3IYcOdKUAIIw34',
-			'11AM': '1qMWjylox_8ewJVT88dm1Cj2c_yiXP2lI',
-			Noon: '1OGt8WGOQC8vOKCo0hCRKicOFMq11IXOi',
-			'1PM': '1d-J_ssFHvUnxLG0Tow36wDkJiRYhJLxN',
-			'2PM': '1hY_poPeeEQ-XuvAXTspcrMViHpVz3a1G',
-			'3PM': '1N2B4EddXyKtGdHewyExBQLFcRXK8ROua',
-			'4PM': '1eOYozFfQu7xiPHQ5SZOqHWh9h_3nB0M8',
-			'5PM': '1oMe23ibjoTXGeoWtuOSiTfCiOM4llf0n',
-			'6PM': '1ZGdSTP7gVBz3nncEWNkN-NU6CbI4JNA7',
-			'7PM': '1iEGZOqdgsN4u70c54OZOKzqCuBYeDEry',
-			'8PM': '1SxiAT0tOQrRWZ4jJT4exvZgMFJmALMlv',
-			'9PM': '1FKlZjtJOqEZ6gbh3w4HJlRxTSCZ0u-Oj',
-			'10PM': '1EC1PEzjfSuTemVH3KcGe6rYu1cuITOE_',
-			'11PM': '1GGk9JsLQRTuAdl39okQYJ9rjlXIidWxd',
-			City: '1NlgBe08MWdbMeAkbDB-XDSFAx-3G7PWI',
-		},
-		Winter: {
-			Midnight: '182578Ih2JO4jFUp-jFd5fqZPR5MnwLDP',
-			'1AM': '13FG3pbKsAvjSK5juN8cGrqbMwhl5GsO8',
-			'2AM': '1oCP7hjGs2GXrrvWn_yMEkmCtYjJR-SGM',
-			'3AM': '1953GE-ncvSwWTirYcqyRJiY2JovNA-sF',
-			'4AM': '1gxvDzVzNzOr2_F1_Zxrguc_kOZd5wrHc',
-			'5AM': '1zvNeEnFEE6N5SXAqj5GnpTgAElqFabyt',
-			'6AM': '1FYEJ13kLEI44GuYS_IiPy2UqL1fs0p8x',
-			'7AM': '1RRLcZ8SX6JfJGn_MuCygw5zs30y0nuXp',
-			'8AM': '1RkWZm_HbW1KF1kFLAvZTgVnSMaE6GJjZ',
-			'9AM': '1nDPD_WwV6ON8dWWXE4eMFe6Ec-1EM2QK',
-			'10AM': '1X8grwEPtvKjI8nuz0UFa1TM0z6g_r6On',
-			'11AM': '1ANBKNW-97DBkVE0kUggVCzW8_5dI10M_',
-			Noon: '1m5prmAbxlAJs05Y2seOpAKEFr7jIpieg',
-			'1PM': '1Ssc9hnrcwJglQZlgM7auo0X10JP_oKqB',
-			'2PM': '1ahg-MwixjEeIeCaD7e7fzf2IT_JG4J-v',
-			'3PM': '1sLaqU_6YVpahj4FSL0QapIip6yvmAgpQ',
-			'4PM': '1RmD_9AzrK8bVSRZXu5C4f8UQQS5FYwh8',
-			'5PM': '1I-1OmmiYj_eyLSeD-W5vwjPY30mFYFuC',
-			'6PM': '1NWGvc42CqWQwuvHsdPPoAj9IEa7AnWzF',
-			'7PM': '1nZ89hX4Cga63lrFBst8v6s783l793wyL',
-			'8PM': '1xeoLU6aAaMkM2ZPRsMRYUeNQ058WfUf7',
-			'9PM': '1dLC4Q9tyyZP7zS1t1Z60lpLlI6_Zjd-M',
-			'10PM': '15SiVKbSvKwzS8B4D_nwrYtQ6OtylF1YB',
-			'11PM': '18QQoFbCOdDVUpEzUJz9IiH6JTjU0aslO',
-			City: '1kog3RcLiOyx2zg5Zna89xyfDiwnmsykM',
-		},
-		City: {
-			Intro: '1jbGbThiTUXmIve7BeKkiklNsahA1UAB3',
-			NewHour: '1tVLwhzwYAeMBBrM6ZYbeBXUMrDKTs7o4',
-			Morning: '19HJ1ljBZeWTQWDwGJUpxrDpWs02ibpoq',
-			Day: '1L8IpC3IvxAgnyI3UEwKcVOhHaMPL-_ml',
-			Evening: '1v4rr5L2t5vgyK2-4jY1DBtUFOAnPU0Xp',
-			Night: '1BpHo4qa0dpo3Ffs7BApflkCa03wYIA2J',
-			Raining: '1NlgBe08MWdbMeAkbDB-XDSFAx-3G7PWI',
-			Winter: '1kog3RcLiOyx2zg5Zna89xyfDiwnmsykM',
-		},
-		Reference: 'https://downloads.khinsider.com/game-soundtracks/album/animal-crossing-city-folk',
-		NewHour: '1tVLwhzwYAeMBBrM6ZYbeBXUMrDKTs7o4',
-	},
-	Credits: 'https://downloads.khinsider.com/',
-	ImportTool: 'https://softgateon.herokuapp.com/directlink/',
-};
-
-export default audioData;
