--- src/components/projects/_actunes/_header.jsx
@@ -1,13 +1,0 @@
-import React from "react";
-
-import banner from "../../../assets/images/AC_App/Animal_Crossing_Logo.png";
-
-const Header = () => {
-  return (
-    <header>
-      <img alt="" src={banner} />
-    </header>
-  );
-};
-
-export default Header;
