--- src/components/projects/_actunes/albums/AlbumNew.jsx
@@ -1,16 +1,0 @@
-import React, { Component } from "react";
-
-class AlbumNew extends Component {
-  constructor() {
-    super();
-    this.state = {
-      //
-    };
-  }
-
-  render() {
-    return <h1>Create New Album</h1>;
-  }
-}
-
-export default AlbumNew;
